# Improves support for concat_dims with size larger than 1