<!-- Thank you very much for your hard work and contribution! -->

The following Pull Request:

- [x] Closes #504 
- [x] Tests are added to demonstrate the new or fixed behavior, and all tests pass
on a local environment.

###
This PR greatly reduces the number of dap responses pydap needs to download in the case multiple (N) urls that aggregate along a dimension `concat_dim` (in most cases, `time`). 
The specific scenario is when the length of the `concat_dim` > 1. In that scenario, `3` dap responses per `concat_dim` per URL are downloaded (Xarray feature when aggregating multiple urls).

This PR makes it so that only `3` dap responses (total!) are downloaded for the `concat_dim` array (e.g. `time`). This is accomplished by forcing the download of all DMR (metadata). This is possible because of the internals of Xarray.

This code assumes that all remote data have the same dimension and identical dim shapes.

### example
Consider a year of MERRA-2 data.
```python
    len(new_urls)
>>> 366
    new_urls[:2] # check the first 2 urls
>>> ['dap4://opendap.earthdata.nasa.gov/collections/C1276812863-GES_DISC/granules/M2T1NXSLV.5.12.4%3AMERRA2_400.tavg1_2d_slv_Nx.20200101.nc4?dap4.ce=/T2M;/U2M;/V2M;/PS;/lon;/time;/lat',
 'dap4://opendap.earthdata.nasa.gov/collections/C1276812863-GES_DISC/granules/M2T1NXSLV.5.12.4%3AMERRA2_400.tavg1_2d_slv_Nx.20200102.nc4?dap4.ce=/T2M;/U2M;/V2M;/PS;/lon;/time;/lat']

## new feature for such collections.
    %% time
    consolidate_metadata(new_urls, concat_dim='time', safe_mode=True, session=cache_session)
>>> datacube has dimensions {'time[0:1:23]', 'lat[0:1:360]', 'lon[0:1:575]'} , and concat dim: `time`
CPU times: user 4.93 s, sys: 1.22 s, total: 6.15 s
Wall time: 1min 2s

### generate xarray dataset
    %%time # jupyter notebook cell magic
    ds = xr.open_mfdataset(new_urls, session=cache_session, engine='pydap', parallel=True, combine='nested', concat_dim='time')
>>> CPU times: user 2.9 s, sys: 498 ms, total: 3.39 s
Wall time: 3.02 s
```
This `O(1) secs` performance is all it will take from now on, when reusing the cache_session.
 
![Screenshot 2025-05-30 at 8 36 10 AM](https://github.com/user-attachments/assets/5a2f4229-4c74-415b-aa12-43317a3ae4c4)



## Old behavior vs New Behavior

With this new behavior the amount of URLs that are downloaded are:
```python
### new feature
    len(cache_session.cache.urls())
>>> 371
```

On the other hand, clearing the `cache_session` and skipping the `consolidate_metadata` step, the timings are:
```python
    %%time
    ds = xr.open_mfdataset(new_urls, session=cache_session, engine='pydap', parallel=True, combine='nested', concat_dim='time')
>>> CPU times: user 21 s, sys: 8.19 s, total: 29.2 s
Wall time: 30min 13s
```
and the total of downloaded URLs are:

```python
    len(cache_session.cache.urls())
>>> 2196 # !!!
```


Overall pydap can now be  10 times faster, because it downloads an order or magnitude fewer downloads