@jgallagher59701 2025-05-30
One question, and this may reveal a fair bit of ignorance on my part, is do you assume that the collection of URLs is structurally homogeneous? Really, it's not the URLs, per se, but the variables being aggregated, but that's a fine point. Is there a way to add a check to ensure that and keep the performance improvement? I cannot see how you would do that, but I thought I'd ask all the same. What does 'safe mode' do?