@Mikejmnez 2025-05-30
> What does 'safe mode' do?

When `safe_mode=True` , all N dmr are downloaded and there is a consistency check that all datasets (associated with each url) have the same dimensions. If this is not True, then program exits. [Here it is](https://github.com/pydap/pydap/blob/265d27e53c5ee54fabc264a0b5fe4330cde7319d/src/pydap/client.py#L232).

`safe_mode=False` by default. However, in this case where the dimension `time` (`concat_dim`) has length >1, `safe_mode` is always set to `True`. That is the case whether the user sets it to `False` or not. THis is implemented in this PR. I found that downloading all DMRs speeds things through in this scenario (mostly because how Xarray's logic operates), and enables me to recycle almost all dap responses.

### making safe_mode even safer
I could add an extra check that all dimension size match across all urls. This is metadata that is already downloaded, and that likely make for a more stringent "safe" mode to enable this performance. A check that indeed, all remote datasets have homogeneous dimensions. Since all dmrs are downloaded, an extra check on Variables is not needed.