@Mikejmnez 2025-05-30
> do you assume that the collection of URLs is structurally homogeneous? Really, it's not the URLs, per se, but the variables being aggregated, but that's a fine point. Is there a way to add a check to ensure that and keep the performance improvement?

Correct, is not just about the URLs. It is all about the `dimensions` (since these are loaded into memory by `Xarray`). `safe_mode=True` essentially performs these checks. `safe_mode=True` is hugely performant in this case. It is what allows the performance gain. 

When length(`concat_dim`)=1, `safe_mode=True` is not as performant compared to `safe_mode=False` (because `safe_mode=False` only downloads `1 DMRs` instead of `N DMRs`), but waaay more performant than the alternative still.